requirements.txt
customtkinter>=5.2.2
Pillow>=12.0.0
ffmpeg-python>=0.2.0
cairosvg>=2.8.2
